Camera execution script and task system
To make the capture process more flexible, the hardcoded preview,
monitor & capture threads need to be moved out of the camera
class. This commit introduces a general purpose camera scheduler
that runs in a background thread monitoring for events. It then
accepts queuing of tasks to execute. Tasks are provided for the
basic operations.