Fix deadlock unmounting camera at startup
The startup code should be using gdk_threads_{enter,leave}
since we make use of threads. When umounting a camera we
entered a recursive main loop. This caused GDK to try to
unlock a mutex that was never locked, corrupting mutex
state. The result was a later hang next time GDK tried to
lock the mutex.

We can't acquire the mutex in the activate callback, since
that can be called from scenarios where the lock is already
held. Thus we move our code into the startup method instead
and just make activate do a window show.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>