Avoid circular updating of controls
When loading config from the camera be careful not to trigger
a change in the UI which in turn triggers a save to the
camera again. Do this by comparing old / new values in the
controls. Also avoiding calling save if no controls have been
marked dirty. Finally also refresh the list of combo box
choices to avoid confusing the UI into triggering an update.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>