Change to use python3 for plugin language
Change the peas plugin loader to use python for the language
instead of gjs. This is to open up possibility to use the
general python library ecosystem which is broader than the
gjs compatible javascript ecosystem. Python3 is chosen instead
of Python2 so we can avoid introducing a problem of having to
switch python versions later.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>