Switch to using a shared library instead of static
When linking to static convenience libraries, the linker will
drop any .o files which are not referenced in the final
executable. This is despite the fact their symbols may be
exported for using by dynamically loaded plugins. Switch to
using a shared library so all code is guaranteed to be kept
available.