Replace progress bar with a spinner
The progress bar in latest GTK does not render in a way
that is attractive in the toolbar. In addition file downloads
are fast enough that detailed progress info is not really
needed. It is sufficient to just use a spinner button to
show the user that an operation is taking place.